Client unary unary interceptor
Implement the unary unary interceptor for the client-side. Interceptors
can be now installed by passing them as a new parameter of the `Channel`
constructor or by giving them as part of the `insecure_channel`
function.

Interceptors are executed within the RPC Asyncio task, receiving as a
response the intercepted call.

The caller has access to all of the intercepted call methods less for
getting the response, which would require to have all of the
interceptors execution finished.